API/requirements.txt
Flask==2.2.2
telnetlib3==2.0.0